+++ Source/CorePatch/CoreUpgrades.cs
@@ -1,7 +1,4 @@
-﻿using RimWorld;
-using Verse;
-
-namespace Replace_Stuff_Compatibility
+﻿namespace Replace_Stuff_Compatibility.CorePatch
 {
 	public class CoreUpgrades: AbstractPatch
 	{