+++ Source/ModPatch/ArmourRacks.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections.Generic;
 using ArmorRacks.ThingComps;
 using ArmorRacks.Things;
-using Replace_Stuff.NewThing;
 using RimWorld;
 using Verse;
 
-namespace Replace_Stuff_Compatibility
+namespace Replace_Stuff_Compatibility.ModPatch
 {
 	public class ArmourRacks : AbstractPatch
 	{