+++ Source/ModPatch/VanillaExpandedFurniture.cs
@@ -1,7 +1,7 @@
 ﻿using RimWorld;
 using Verse;
 
-namespace Replace_Stuff_Compatibility
+namespace Replace_Stuff_Compatibility.ModPatch
 {
 	public class VanillaExpandedFurniture: AbstractPatch
 	{